pyproject.toml
[project]
name = "calories_prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pytest>=8.4.2",
    "torch>=2.9.0",
    "scikit-learn>=1.6.0",
    "seaborn>=0.13.0",
]
